+++ RELEASE_NOTES.md
@@ -1,4 +1,7 @@
-﻿# Hotspot Manager v0.5.16
+﻿# Hotspot Manager v0.5.17
 
 ## Yenilikler
 
+
+## Notlar
+- Optimized startup speed (Task Scheduler + High Priority) and added Single Instance Lock